030_fractions: add initial version of benchmark
This benchmark has been developed to test whether or not we can reason
effectively about rational numbers without requiring an explicit
fraction type in the language itself.  The benchmark is pretty basic in
many respects, but hopefully it will demonstrate the issue.

At the moment, unfortunately, the benchmark does not compile because of
problems handling type invariants.