Update build system to using correct working dirs
Previously, the build system was not using the correct working directory
when compiling each benchmark.  This is now fixed, meaning that it is
properly trying to compile them.  However, the next problem to be fixed
is that it can't find the standard library.  In order to make this work
we need the (new) Ant Task to support the whileypath attribute.