Update all brenchmarks to compile; or be "broken"
This fixes a number of relatively minor problems related to changes or
simplifications in the syntax of Whiley.  In many cases, these
simplifications will be returned to their former glory at some point.
But, for now, the goal is to get these benchmarks compiling!!

At this point, there remain a number of benchmarks which are no longer
compiler.  These have been placed into the broken/ directory to ensure
they are not included during compilation.

Finally, this version compiles against v0.4.2-SNAPSHOT.  Unfortunately,
this means that it won't pass CI because that version of wyc is not yet
released!!  argh.  Oh well.