028_flag: updated specification and comments
Whiley is now capable of verifying a more powerful loop invariant for
the dutch national flag problem.  The current specification is now
pretty good, though it does not ensure the same counts of elements in
the resulting array.

In addition, this updates the documentation to be a little nicer.
Overall, this benchmark is looking pretty sweet now!!