010_sort: fix for broken precondition
The precondition for io::println requires that an ASCII string is provided.
However, in some places, a raw integer array was being provided instead.  The
fix for this was to add a function ascii::to_string for converting an integer
array into a string and use this at the necessary points.