010_sort: fix for broken postcondition
The return type invariant for sort specified (incorrectly) that the whole array
returned must be sorted.  In fact, only the portion of the array examined should
be sorted, not the whole array.  This was resolved by adding a new property,
sorted, to express that part of an array was sorted.010_sort: fix for broken
postcondition

The return type invariant for sort specified (incorrectly) that the whole array
returned must be sorted.  In fact, only the portion of the array examined should
be sorted, not the whole array.  This was resolved by adding a new property,
sorted, to express that part of an array was sorted.010_sort: fix for broken
postcondition

The return type invariant for sort specified (incorrectly) that the whole array
returned must be sorted.  In fact, only the portion of the array examined should
be sorted, not the whole array.  This was resolved by adding a new property,
sorted, to express that part of an array was sorted.010_sort: fix for broken
postcondition

The return type invariant for sort specified (incorrectly) that the whole array
returned must be sorted.  In fact, only the portion of the array examined should
be sorted, not the whole array.  This was resolved by adding a new property,
sorted, to express that part of an array was sorted.010_sort: fix for broken
postcondition

The return type invariant for sort specified (incorrectly) that the whole array
returned must be sorted.  In fact, only the portion of the array examined should
be sorted, not the whole array.  This was resolved by adding a new property,
sorted, to express that part of an array was sorted.010_sort: fix for broken
postcondition

The return type invariant for sort specified (incorrectly) that the whole array
returned must be sorted.  In fact, only the portion of the array examined should
be sorted, not the whole array.  This was resolved by adding a new property,
sorted, to express that part of an array was sorted.010_sort: fix for broken
postcondition

The return type invariant for sort specified (incorrectly) that the whole array
returned must be sorted.  In fact, only the portion of the array examined should
be sorted, not the whole array.  This was resolved by adding a new property,
sorted, to express that part of an array was sorted.010_sort: fix for broken
postcondition

The return type invariant for sort specified (incorrectly) that the whole array
returned must be sorted.  In fact, only the portion of the array examined should
be sorted, not the whole array.  This was resolved by adding a new property,
sorted, to express that part of an array was sorted.010_sort: fix for broken
postcondition

The return type invariant for sort specified (incorrectly) that the whole array
returned must be sorted.  In fact, only the portion of the array examined should
be sorted, not the whole array.  This was resolved by adding a new property,
sorted, to express that part of an array was sorted.010_sort: fix for broken
postcondition

The return type invariant for sort specified (incorrectly) that the whole array
returned must be sorted.  In fact, only the portion of the array examined should
be sorted, not the whole array.  This was resolved by adding a new property,
sorted, to express that part of an array was sorted.