012_cyclic: fix for broken type invariant
There was a broken type invariant being triggered by QuickCheck.  The issue was
a temporary breakage of the buffer state whilst the write (resp. read) pointers
were being updated.  As such, it wasn't a true bug per se and the fix was simply
to assign to a temporary variable with a less restrictive type invariant.