025_tries: fix for broken precondition
There was a failing precondition which indicating a real bug related to the way
that the trie.size field was updated after a new transition was added.  This was
fixed easily enough by modifying the code for calculating the updated size.

In addition, pushed through some simpliciations by replacing code with calls to
the standard library.  Also fleshed out the specifications a bit more for the
add(Trie,Transition) method.