027_c_string: fix for broken type invariant
There was a broken type invariant for a parameter to the strcpy method.  The
problem was that it required a full C_String when, in fact, any old byte array
would do.  That's because the copy operation *creates* a C string.  Therefore,
have relatex this type invariant.

In addition, observed another bug with strcpy which was not copying the NULL
terminator over into the new string.  This wasn't spotted by QuickCheck because
there was no violation of the specification.  To catch it, an assertion was
added.  In fact, what would be useful, would be some mechanism for describing
before and after states for references.