Still working on Z3 regression
It would be nice to boil it down to a simple example.  I think this
could be done with the extend function.  That is, we should be able to
show properties being implied from the generic postcondition of append.

In the Dafny version, they boil this down to two lemmas.